Phase3_Bootstrap/requirements.txt
# ============================================================================
# Phase 3 Bootstrap — Python dependencies
# Pinned to major versions that were verified during Phase 1/2.
# ============================================================================

# ---- Interactive UI ----
rich>=13.7,<14.0                 # pretty printing, progress bars, tables
questionary>=2.0,<3.0            # interactive prompts (arrow-key menus)
prompt_toolkit>=3.0,<4.0         # underlying prompt engine

# ---- Configuration ----
pydantic>=2.5,<3.0               # schema validation
PyYAML>=6.0,<7.0                 # config files
python-dotenv>=1.0,<2.0          # .env parsing
jinja2>=3.1,<4.0                 # template rendering

# ---- Google Cloud ----
google-cloud-storage>=2.14,<4.0
google-cloud-secret-manager>=2.18,<3.0
google-cloud-discoveryengine>=0.11,<1.0
google-cloud-resource-manager>=1.12,<2.0
google-cloud-billing>=1.12,<2.0
google-cloud-logging>=3.9,<4.0
google-api-python-client>=2.100,<3.0
google-auth>=2.23,<3.0
google-auth-oauthlib>=1.1,<2.0

# ---- HTTP ----
requests>=2.31,<3.0
httpx>=0.25,<1.0

# ---- Dev / test ----
pytest>=7.4,<9.0
pytest-mock>=3.12,<4.0
